masterthesis/requirements.txt
argparse
# boto3
# botocore
matplotlib
numpy
pandas
pathlib
pillow
pycocotools
# pytorch_lightning
scikit-learn
setuptools
torch>=1.6.0
torchvision>=0.5.0
tqdm
ipython
opencv-python
colormap
